Move pcluster delete command into cli module
